Added support for 16-bit logical bus in FRU write
This commit adds support for 16-bit logical bus. It fixes the
issue where FRU write operation failed when the I2C bus number
exceeded 255.

Tested:
Fru read and write commands working fine.
Fru functionality verified.

Note:
Due to hardware unavailability not able to test
the I2C bus number exceeded 255

+++ src/fru_device/fru_utils.cpp
@@ -1646,7 +1646,7 @@ std::optional<int> findIndexForFRU(
 std::optional<std::string> getProductName(
     std::vector<uint8_t>& device,
     std::flat_map<std::string, std::string, std::less<>>& formattedFRU,
-    uint32_t bus, uint32_t address, size_t& unknownBusObjectCount)
+    uint16_t bus, uint8_t address, size_t& unknownBusObjectCount)
 {
     std::string productName;
 
@@ -1686,12 +1686,11 @@ std::optional<std::string> getProductName(
     return productName;
 }
 
-bool getFruData(std::vector<uint8_t>& fruData, uint32_t bus, uint32_t address)
+bool getFruData(std::vector<uint8_t>& fruData, uint16_t bus, uint8_t address)
 {
     try
     {
-        fruData = getFRUInfo(static_cast<uint16_t>(bus),
-                             static_cast<uint8_t>(address));
+        fruData = getFRUInfo(bus, address);
     }
     catch (const std::invalid_argument& e)
     {
